move logger out of utils
