+++ packages/web/src/features/git/logger.ts
@@ -1,0 +1,3 @@
+import { createLogger } from "@sourcebot/shared";
+
+export const logger = createLogger('git');
